ShippingRateInfo component, separate ShippingZoneInfo, cleanups
