+++ src/shipping/components/ShippingRateInfo/ShippingRateInfo.tsx
@@ -1,0 +1,52 @@
+import Card from "@material-ui/core/Card";
+import CardContent from "@material-ui/core/CardContent";
+import TextField from "@material-ui/core/TextField";
+import CardTitle from "@saleor/components/CardTitle";
+import { ShippingErrorFragment } from "@saleor/fragments/types/ShippingErrorFragment";
+import { commonMessages } from "@saleor/intl";
+import { getFormErrors } from "@saleor/utils/errors";
+import getShippingErrorMessage from "@saleor/utils/errors/shipping";
+import React from "react";
+import { useIntl } from "react-intl";
+
+export interface ShippingRateInfoProps {
+  data: Record<"name", string>;
+  disabled: boolean;
+  errors: ShippingErrorFragment[];
+  onChange: (event: React.ChangeEvent<any>) => void;
+}
+
+const ShippingRateInfo: React.FC<ShippingRateInfoProps> = ({
+  data,
+  disabled,
+  errors,
+  onChange
+}) => {
+  const intl = useIntl();
+
+  const formErrors = getFormErrors(["name"], errors);
+
+  return (
+    <Card>
+      <CardTitle
+        title={intl.formatMessage(commonMessages.generalInformations)}
+      />
+      <CardContent>
+        <TextField
+          disabled={disabled}
+          error={!!formErrors.name}
+          fullWidth
+          helperText={getShippingErrorMessage(formErrors.name, intl)}
+          label={intl.formatMessage({
+            defaultMessage: "Shipping rate name"
+          })}
+          name="name"
+          value={data.name}
+          onChange={onChange}
+        />
+      </CardContent>
+    </Card>
+  );
+};
+ShippingRateInfo.displayName = "ShippingRateInfo";
+export default ShippingRateInfo;
