+++ src/shipping/components/ShippingRateInfo/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./ShippingRateInfo";
+export * from "./ShippingRateInfo";
