+++ src/shipping/components/ShippingZoneCreatePage/ShippingZoneCreatePage.tsx
@@ -1,5 +1,3 @@
-import makeStyles from "@material-ui/core/styles/makeStyles";
-import TextField from "@material-ui/core/TextField";
 import AppHeader from "@saleor/components/AppHeader";
 import CardSpacer from "@saleor/components/CardSpacer";
 import { ConfirmButtonTransitionState } from "@saleor/components/ConfirmButton";
@@ -13,7 +11,7 @@ import { CountryFragment } from "@saleor/fragments/types/CountryFragment";
 import { ShippingErrorFragment } from "@saleor/fragments/types/ShippingErrorFragment";
 import { sectionNames } from "@saleor/intl";
 import React from "react";
-import { defineMessages, FormattedMessage, useIntl } from "react-intl";
+import { defineMessages, useIntl } from "react-intl";
 
 import ShippingZoneCountriesAssignDialog from "../ShippingZoneCountriesAssignDialog";
 import ShippingZoneInfo from "../ShippingZoneInfo";
@@ -25,53 +23,22 @@ export interface FormData {
   name: string;
 }
 
-const MAX_DESCRIPTION_LENGTH = 300;
-
-const useStyles = makeStyles(
-  {
-    label: {
-      flex: 1
-    },
-    labelContainer: {
-      "& span": {
-        paddingRight: 30
-      },
-      display: "flex"
-    }
-  },
-  { name: "ShippingZoneCreatePage" }
-);
-
 const messages = defineMessages({
   countries: {
     defaultMessage: "Countries",
-    description: "countries"
+    description: "country list header"
   },
   createZone: {
     defaultMessage: "Create New Shipping Zone",
-    description: "header"
+    description: "section header"
   },
   defaultZone: {
     defaultMessage:
-      "This is default shipping zone, which means that it covers all of the countries which are not assigned to other shipping zones",
-    description: "default shipping zone"
-  },
-  description: {
-    defaultMessage: "Description",
-    description: "description"
-  },
-  descriptionMessage: {
-    defaultMessage: "Description of a shipping zone.",
-    description: "field placeholder"
+      "This is default shipping zone, which means that it covers all of the countries which are not assigned to other shipping zones"
   },
   noCountriesAssigned: {
     defaultMessage:
-      "Currently, there are no countries assigned to this shipping zone",
-    description: "no countries assigned to zone"
-  },
-  shipping: {
-    defaultMessage: "Shipping",
-    description: "shipping"
+      "Currently, there are no countries assigned to this shipping zone"
   }
 });
 
@@ -93,7 +60,6 @@ const ShippingZoneCreatePage: React.FC<ShippingZoneCreatePageProps> = ({
   saveButtonBarState
 }) => {
   const intl = useIntl();
-  const classes = useStyles({});
   const [isModalOpened, setModalStatus] = React.useState(false);
   const toggleModal = () => setModalStatus(!isModalOpened);
 
@@ -120,42 +86,6 @@ const ShippingZoneCreatePage: React.FC<ShippingZoneCreatePageProps> = ({
                   disabled={disabled}
                   errors={errors}
                   onChange={change}
-                />
-                <CardSpacer />
-                <TextField
-                  error={data.description.length > MAX_DESCRIPTION_LENGTH}
-                  name={"description"}
-                  label={
-                    <div className={classes.labelContainer}>
-                      <div className={classes.label}>
-                        <FormattedMessage {...messages.description} />
-                      </div>
-                      {data.description?.length > 0 && (
-                        <span>
-                          <FormattedMessage
-                            defaultMessage="{numberOfCharacters} of {maxCharacters} characters"
-                            description="character limit"
-                            values={{
-                              maxCharacters: MAX_DESCRIPTION_LENGTH,
-                              numberOfCharacters: data.description.length
-                            }}
-                          />
-                        </span>
-                      )}
-                    </div>
-                  }
-                  InputProps={{
-                    inputProps: {
-                      maxLength: MAX_DESCRIPTION_LENGTH
-                    }
-                  }}
-                  value={data.description}
-                  onChange={change}
-                  disabled={disabled}
-                  fullWidth
-                  multiline
-                  placeholder={intl.formatMessage(messages.descriptionMessage)}
-                  rows={10}
                 />
                 <CardSpacer />
                 <CountryList
